+++ constraint_bodies.cpp
@@ -9,7 +9,7 @@
 // Rigid bar 1 methods
 // This rigid bar attaches at a pivot at point0, and a mass at point1
 
-float Rigid_Bar_1::constraint(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_1::constraint(Circular_Rigid_Body *mass_list)
 {
 	// Given a state vector q, return a scalar value of this constraint function.
 	// For single attachments, where one end is fixed. Connects a maximum of two masses
@@ -18,43 +18,43 @@ float Rigid_Bar_1::constraint(Circular_Rigid_Body *mass_list)
 	// First constraint equation
 	// Set the pivot to the origin
 	// Shift x1 and y1 so they are relative to the pivot
-	float x1 = mass_list[attached_mass].pos.x;
-	float y1 = mass_list[attached_mass].pos.y;
+	double x1 = mass_list[attached_mass].pos.x;
+	double y1 = mass_list[attached_mass].pos.y;
 
 	// Calculate the constraint
-	float length = vector_magnitude({initial_point.x - pivot.x, initial_point.y - pivot.y});
+	double length = vector_magnitude({initial_point.x - pivot.x, initial_point.y - pivot.y});
 	return ((x1 - pivot.x) * (x1 - pivot.x)) + ((y1 - pivot.y) * (y1 - pivot.y)) - (length * length);
 }
 
-float Rigid_Bar_1::constraint_time_derivative(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_1::constraint_time_derivative(Circular_Rigid_Body *mass_list)
 {
 	// Returns a value for the constraint time derivative
-	float x1 = mass_list[attached_mass].pos.x;
-	float y1 = mass_list[attached_mass].pos.y;
-	float x1_dot = mass_list[attached_mass].linear_vel.x;
-	float y1_dot = mass_list[attached_mass].linear_vel.y;
+	double x1 = mass_list[attached_mass].pos.x;
+	double y1 = mass_list[attached_mass].pos.y;
+	double x1_dot = mass_list[attached_mass].linear_vel.x;
+	double y1_dot = mass_list[attached_mass].linear_vel.y;
 	return 2.0 * x1_dot * (x1 - pivot.x) + 2.0 * y1_dot * (y1 - pivot.y);
 
 }
 
 // Partial of constraint function with respect to x
-float Rigid_Bar_1::jacobian_entry_x(Circular_Rigid_Body* mass_list)
+double Rigid_Bar_1::jacobian_entry_x(Circular_Rigid_Body* mass_list)
 {
 	return 2.0 * (mass_list[attached_mass].pos.x - pivot.x);
 }
 
 // Partial of constraint function with respect to y
-float Rigid_Bar_1::jacobian_entry_y(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_1::jacobian_entry_y(Circular_Rigid_Body *mass_list)
 {
 	return 2.0 * (mass_list[attached_mass].pos.y - pivot.y);
 }
 // partial of constraint time derivative wrt x and y
-float Rigid_Bar_1::jacobian_derivative_entry_x(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_1::jacobian_derivative_entry_x(Circular_Rigid_Body *mass_list)
 {
 	return 2.0 * mass_list[attached_mass].linear_vel.x;
 }
 
-float Rigid_Bar_1::jacobian_derivative_entry_y(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_1::jacobian_derivative_entry_y(Circular_Rigid_Body *mass_list)
 {
 	return 2.0 * mass_list[attached_mass].linear_vel.y;
 }
@@ -81,93 +81,93 @@ void Rigid_Bar_1::draw_to_screen(Circular_Rigid_Body *mass_list, void *first_pix
 // Rigid bar 2 functions
 // Attaches to two moving mass objects
 
-float Rigid_Bar_2::constraint(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_2::constraint(Circular_Rigid_Body *mass_list)
 {
 
-		float x1 = mass_list[attached_masses[0]].pos.x;
-		float y1 = mass_list[attached_masses[0]].pos.y;
-		float x2 = mass_list[attached_masses[1]].pos.x;
-		float y2 = mass_list[attached_masses[1]].pos.y;
+		double x1 = mass_list[attached_masses[0]].pos.x;
+		double y1 = mass_list[attached_masses[0]].pos.y;
+		double x2 = mass_list[attached_masses[1]].pos.x;
+		double y2 = mass_list[attached_masses[1]].pos.y;
 
 		// Calculate the constraint
-		float length = vector_magnitude({initial_point1.x - initial_point0.x, initial_point1.y - initial_point0.y});
+		double length = vector_magnitude({initial_point1.x - initial_point0.x, initial_point1.y - initial_point0.y});
 		return ((x2 - x1) * (x2 - x1)) + ((y2 - y1) * (y2 - y1)) - (length * length);
 }
 
-float Rigid_Bar_2::constraint_time_derivative(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_2::constraint_time_derivative(Circular_Rigid_Body *mass_list)
 {
-	float x1 = mass_list[attached_masses[0]].pos.x;
-	float y1 = mass_list[attached_masses[0]].pos.y;
-	float x2 = mass_list[attached_masses[1]].pos.x;
-	float y2 = mass_list[attached_masses[1]].pos.y;
+	double x1 = mass_list[attached_masses[0]].pos.x;
+	double y1 = mass_list[attached_masses[0]].pos.y;
+	double x2 = mass_list[attached_masses[1]].pos.x;
+	double y2 = mass_list[attached_masses[1]].pos.y;
 
-	float x1_dot = mass_list[attached_masses[0]].linear_vel.x;
-	float y1_dot = mass_list[attached_masses[0]].linear_vel.y;
-	float x2_dot = mass_list[attached_masses[1]].linear_vel.x;
-	float y2_dot = mass_list[attached_masses[1]].linear_vel.y;
+	double x1_dot = mass_list[attached_masses[0]].linear_vel.x;
+	double y1_dot = mass_list[attached_masses[0]].linear_vel.y;
+	double x2_dot = mass_list[attached_masses[1]].linear_vel.x;
+	double y2_dot = mass_list[attached_masses[1]].linear_vel.y;
 
 	return 2.0 * x1_dot * (x1 - x2) + 2.0 * y1_dot * (y1 - y2) + 2.0 * x2_dot * (x2 - x1) + 2.0 * y2_dot * (y2 - y1);
 }
 
-float Rigid_Bar_2::jacobian_entry_x1(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_2::jacobian_entry_x1(Circular_Rigid_Body *mass_list)
 {
 
-	float x1 = mass_list[attached_masses[0]].pos.x;
-	float x2 = mass_list[attached_masses[1]].pos.x;
+	double x1 = mass_list[attached_masses[0]].pos.x;
+	double x2 = mass_list[attached_masses[1]].pos.x;
 	return -2.0 * (x2 - x1);
 
 }
 
-float Rigid_Bar_2::jacobian_entry_y1(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_2::jacobian_entry_y1(Circular_Rigid_Body *mass_list)
 {
 
-	float y1 = mass_list[attached_masses[0]].pos.y;
-	float y2 = mass_list[attached_masses[1]].pos.y;
+	double y1 = mass_list[attached_masses[0]].pos.y;
+	double y2 = mass_list[attached_masses[1]].pos.y;
 	return -2.0 * (y2 - y1);
 }
 
-float Rigid_Bar_2::jacobian_entry_x2(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_2::jacobian_entry_x2(Circular_Rigid_Body *mass_list)
 {
 
-	float x1 = mass_list[attached_masses[0]].pos.x;
-	float x2 = mass_list[attached_masses[1]].pos.x;
+	double x1 = mass_list[attached_masses[0]].pos.x;
+	double x2 = mass_list[attached_masses[1]].pos.x;
 	return 2.0 * (x2 - x1);
 }
 
-float Rigid_Bar_2::jacobian_entry_y2(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_2::jacobian_entry_y2(Circular_Rigid_Body *mass_list)
 {
 
-	float y1 = mass_list[attached_masses[0]].pos.y;
-	float y2 = mass_list[attached_masses[1]].pos.y;
+	double y1 = mass_list[attached_masses[0]].pos.y;
+	double y2 = mass_list[attached_masses[1]].pos.y;
 	return 2.0 * (y2 - y1);
 }
 
 // partial of constraint time derivative wrt x1, y2, x2, y2
-float Rigid_Bar_2::jacobian_derivative_entry_x1(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_2::jacobian_derivative_entry_x1(Circular_Rigid_Body *mass_list)
 {
-	float x1_dot = mass_list[attached_masses[0]].linear_vel.x;
-	float x2_dot = mass_list[attached_masses[1]].linear_vel.x;
+	double x1_dot = mass_list[attached_masses[0]].linear_vel.x;
+	double x2_dot = mass_list[attached_masses[1]].linear_vel.x;
 	return 2.0 * (x1_dot - x2_dot);
 }
 
-float Rigid_Bar_2::jacobian_derivative_entry_y1(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_2::jacobian_derivative_entry_y1(Circular_Rigid_Body *mass_list)
 {
-	float y1_dot = mass_list[attached_masses[0]].linear_vel.y;
-	float y2_dot = mass_list[attached_masses[1]].linear_vel.y;
+	double y1_dot = mass_list[attached_masses[0]].linear_vel.y;
+	double y2_dot = mass_list[attached_masses[1]].linear_vel.y;
 	return 2.0 * (y1_dot - y2_dot);
 }
 
-float Rigid_Bar_2::jacobian_derivative_entry_x2(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_2::jacobian_derivative_entry_x2(Circular_Rigid_Body *mass_list)
 {
-	float x1_dot = mass_list[attached_masses[0]].linear_vel.x;
-	float x2_dot = mass_list[attached_masses[1]].linear_vel.x;
+	double x1_dot = mass_list[attached_masses[0]].linear_vel.x;
+	double x2_dot = mass_list[attached_masses[1]].linear_vel.x;
 	return 2.0 * (x2_dot - x1_dot);
 }
 
-float Rigid_Bar_2::jacobian_derivative_entry_y2(Circular_Rigid_Body *mass_list)
+double Rigid_Bar_2::jacobian_derivative_entry_y2(Circular_Rigid_Body *mass_list)
 {
-	float y1_dot = mass_list[attached_masses[0]].linear_vel.y;
-	float y2_dot = mass_list[attached_masses[1]].linear_vel.y;
+	double y1_dot = mass_list[attached_masses[0]].linear_vel.y;
+	double y2_dot = mass_list[attached_masses[1]].linear_vel.y;
 	return 2.0 * (y2_dot - y1_dot);
 }
