+++ constraint_bodies.h
@@ -25,16 +25,16 @@ class Rigid_Bar_1
 	int thickness;
 
 	// Constraint equation of rigid rod
-	float constraint(Circular_Rigid_Body *mass_list);
-	float constraint_time_derivative(Circular_Rigid_Body *mass_list);
+	double constraint(Circular_Rigid_Body *mass_list);
+	double constraint_time_derivative(Circular_Rigid_Body *mass_list);
 	// Constraint derivatives
 	// Partial derivatives wrt x and y
-	float jacobian_entry_x(Circular_Rigid_Body* mass_list);
-	float jacobian_entry_y(Circular_Rigid_Body *mass_list);
+	double jacobian_entry_x(Circular_Rigid_Body* mass_list);
+	double jacobian_entry_y(Circular_Rigid_Body *mass_list);
 
 	// Partial of constraint time derivative wrt x and y.
-	float jacobian_derivative_entry_x(Circular_Rigid_Body *mass_list);
-	float jacobian_derivative_entry_y(Circular_Rigid_Body *mass_list);
+	double jacobian_derivative_entry_x(Circular_Rigid_Body *mass_list);
+	double jacobian_derivative_entry_y(Circular_Rigid_Body *mass_list);
 
 	void determine_initial_point(Circular_Rigid_Body *mass_list);
 
@@ -62,20 +62,20 @@ class Rigid_Bar_2
 	int thickness;
 
 	// Constraint equation of rigid rod connecting two moving masses
-	float constraint(Circular_Rigid_Body *mass_list);
-	float constraint_time_derivative(Circular_Rigid_Body *mass_list);
+	double constraint(Circular_Rigid_Body *mass_list);
+	double constraint_time_derivative(Circular_Rigid_Body *mass_list);
 
 	// Partial of constraint function wrt x1, y1, x2, y2
-	float jacobian_entry_x1(Circular_Rigid_Body *mass_list);
-	float jacobian_entry_y1(Circular_Rigid_Body *mass_list);
-	float jacobian_entry_x2(Circular_Rigid_Body *mass_list);
-	float jacobian_entry_y2(Circular_Rigid_Body *mass_list);
+	double jacobian_entry_x1(Circular_Rigid_Body *mass_list);
+	double jacobian_entry_y1(Circular_Rigid_Body *mass_list);
+	double jacobian_entry_x2(Circular_Rigid_Body *mass_list);
+	double jacobian_entry_y2(Circular_Rigid_Body *mass_list);
 
 	// partial of constraint time derivative wrt x1, y1, x2, y2
-	float jacobian_derivative_entry_x1(Circular_Rigid_Body *mass_list);
-	float jacobian_derivative_entry_y1(Circular_Rigid_Body *mass_list);
-	float jacobian_derivative_entry_x2(Circular_Rigid_Body *mass_list);
-	float jacobian_derivative_entry_y2(Circular_Rigid_Body *mass_list);
+	double jacobian_derivative_entry_x1(Circular_Rigid_Body *mass_list);
+	double jacobian_derivative_entry_y1(Circular_Rigid_Body *mass_list);
+	double jacobian_derivative_entry_x2(Circular_Rigid_Body *mass_list);
+	double jacobian_derivative_entry_y2(Circular_Rigid_Body *mass_list);
 
 	void determine_initial_points(Circular_Rigid_Body *mass_list);
