+++ vectors.cpp
@@ -6,7 +6,7 @@
 // List of vector functions
 
 // Magnitude
-float vector_magnitude(Vector a)
+double vector_magnitude(Vector a)
 {
 	return sqrt((a.x * a.x) + (a.y * a.y));
 }
@@ -15,7 +15,7 @@ float vector_magnitude(Vector a)
 Vector vector_direction(Vector a)
 {
 	Vector unit;
-	float mag = vector_magnitude(a);
+	double mag = vector_magnitude(a);
 
 	unit.x = a.x / mag;
 	unit.y = a.y / mag;
@@ -34,14 +34,14 @@ Vector vector_addition(Vector a, Vector b)
 }
 
 // Dot product: Returns a scaler
-float vector_dp(Vector a, Vector b)
+double vector_dp(Vector a, Vector b)
 {
 	return a.x * b.x + a.y * b.y;
 }
 
 // Angle between two vectors: Dot product method
-float angle_between_vectors(Vector a, Vector b)
+double angle_between_vectors(Vector a, Vector b)
 {
-	float angle = acos(vector_dp(a, b) / (vector_magnitude(a) * vector_magnitude(b)));
+	double angle = (double)acos(vector_dp(a, b) / (vector_magnitude(a) * vector_magnitude(b)));
 	return angle;
 }
